Fix build issue with Linux v6.4 and later
Remove argument `THIS_MODULE` when calling class_create() if the kernel
version is greater than or equal to v6.4.

In Linux Kernel v6.4, the API for class_create() was updated, reducing
the number of parameters from 2 to 1. Due to this change, the original
function call to class_create() will cause a 'too many arguments' error
when building the kernel module.