Use framebuffer_alloc() to initialize fb_info
The reason to do such changes is that the flag FBINFO_DEFAULT has
been removed from the Linux kernel since v6.6, which was the default
value of fb_info.flags.

To initialize the fb_info structure (and the value of fb_info.flags)
correctly without using FBINFO_DEFAULT, the framebuffer_alloc()
function provided by the Linux kernel should be used.

Since framebuffer_alloc() only accepts pointer to fb_info, the type of
the info field in struct vcamfb_info was changed from
'struct fb_info' to 'pointer to struct fb_info'.

As a result, framebuffer_release() was added to deallocate the memory
pointed to by the info field. Additionally, several changes were made
to accommodate the change of the info field's type.